creation de d'une librairie pour appliquer les fonctions de communication avec la base de donnée
